🧩 5NHXTN task: refresh task artifacts after commit
