+++ .agentplane/tasks/202605171737-5NHXTN/pr/meta.json
@@ -3,11 +3,14 @@
   "branch": "task/202605171737-5NHXTN/fix-3843-base-start-ready",
   "created_at": "2026-05-17T17:37:29.120Z",
   "head_sha": "a273f138da1f0ad93bb22c72560ae0741842ea0e",
-  "last_verified_at": "2026-05-17T17:47:08.756Z",
-  "last_verified_sha": "3e171077adaf143f9e060aeb285bb1e61fd0e026",
+  "last_verified_at": "2026-05-17T17:54:39.890Z",
+  "last_verified_sha": "a273f138da1f0ad93bb22c72560ae0741842ea0e",
+  "pr_number": 3850,
+  "pr_url": "https://github.com/basilisk-labs/agentplane/pull/3850",
   "schema_version": 1,
+  "status": "OPEN",
   "task_id": "202605171737-5NHXTN",
-  "updated_at": "2026-05-17T17:48:36.638Z",
+  "updated_at": "2026-05-17T17:49:27.196Z",
   "verify": {
     "status": "pass"
   }
