+++ .agentplane/tasks/202605171737-5NHXTN/pr/review.md
@@ -13,7 +13,7 @@ Created: 2026-05-17T17:37:29.120Z
 ## Verification
 
 - State: ok
-- Note: Command: bun test packages/agentplane/src/cli/run-cli.core.pr-flow.test.ts --runInBand; Result: pass; Evidence: 17 tests passed including branch_pr base start-ready recovery regression. Scope: work start/task start-ready route behavior. Command: bunx vitest --config vitest.workspace.ts run --project agentplane packages/agentplane/src/commands/task/start.unit.test.ts packages/agentplane/src/commands/task/finish.state.unit.test.ts; Result: pass; Evidence: 2 files, 9 tests passed. Scope: existing task start/finish lifecycle unit coverage. Command: bun run typecheck; Result: pass; Evidence: tsc -b exited 0. Scope: repo TypeScript project references. Command: bun run format:changed; Result: pass; Evidence: all matched files use Prettier style. Scope: touched files. Command: bun run lint:core -- packages/agentplane/src/commands/task/start-ready.ts packages/agentplane/src/cli/run-cli.core.pr-flow.test.ts; Result: pass; Evidence: eslint exited 0. Scope: repo core lint. Command: node .agentplane/policy/check-routing.mjs; Result: pass; Evidence: policy routing OK. Scope: policy routing.
+- Note: Command: bun test packages/agentplane/src/cli/run-cli.core.pr-flow.test.ts packages/agentplane/src/cli/run-cli.core.pr-flow.start-ready.test.ts --runInBand; Result: pass; Evidence: 17 tests passed across branch_pr work start/start-ready routing suites, including base checkout recovery regression. Scope: work start/task start-ready route behavior. Command: bun test packages/agentplane/src/cli/run-cli.core.pr-flow.start-ready.test.ts --runInBand; Result: pass; Evidence: 1 focused regression test passed. Scope: issue #3843 route. Command: bunx vitest --config vitest.workspace.ts run --project agentplane packages/agentplane/src/commands/task/start.unit.test.ts packages/agentplane/src/commands/task/finish.state.unit.test.ts; Result: pass; Evidence: 2 files, 9 tests passed. Scope: existing task start/finish lifecycle unit coverage. Command: bun run typecheck; Result: pass; Evidence: tsc -b exited 0. Scope: repo TypeScript project references. Command: bun run format:changed; Result: pass; Evidence: all matched files use Prettier style. Scope: touched files. Command: bun run hotspots:check; Result: pass; Evidence: hotspot threshold check passed after moving regression out of oversized pr-flow suite. Scope: CI hotspot gate. Command: bun run lint:core -- packages/agentplane/src/commands/task/start-ready.ts packages/agentplane/src/cli/run-cli.core.pr-flow.test.ts; Result: pass; Evidence: eslint exited 0. Scope: repo core lint. Command: node .agentplane/policy/check-routing.mjs; Result: pass; Evidence: policy routing OK. Scope: policy routing.
 - Canonical workflow state lives in the task README.
 
 ## Handoff Notes
@@ -24,7 +24,7 @@ Created: 2026-05-17T17:37:29.120Z
 <details>
 <summary>Raw evidence</summary>
 
-- Updated: 2026-05-17T17:48:36.638Z
+- Updated: 2026-05-17T17:49:27.196Z
 - Branch: task/202605171737-5NHXTN/fix-3843-base-start-ready
 - Head: a273f138da1f
